Move methods from modules into single file, ErrorScanner class
